// src/utils/gmx/context/SyntheticsEvents/types.ts
export type PendingFundingFeeSettlementData = {
  orders: PendingOrderData[];
  positions: PendingPositionUpdate[];
};

export type OrderCreatedEventData = {
  key: string;
  account: string;
  receiver: string;
  callbackContract: string;
  marketAddress: string;
  initialCollateralTokenAddress: string;
  swapPath: string[];
  sizeDeltaUsd: BigNumber;
  initialCollateralDeltaAmount: BigNumber;
  contractTriggerPrice: BigNumber;
  contractAcceptablePrice: BigNumber;
  executionFee: BigNumber;
  callbackGasLimit: BigNumber;
  minOutputAmount: BigNumber;
  updatedAtBlock: BigNumber;
  orderType: OrderType;
  isLong: boolean;
  shouldUnwrapNativeToken: boolean;
  isFrozen: boolean;
};

export type PendingOrderData = {
  account: string;
  marketAddress: string;
  initialCollateralTokenAddress: string;
  swapPath: string[];
  initialCollateralDeltaAmount: BigNumber;
  minOutputAmount: BigNumber;
  sizeDeltaUsd: BigNumber;
  isLong: boolean;
  shouldUnwrapNativeToken: boolean;
  orderType: OrderType;
};
// ...
export type MultiTransactionStatus<TEventData> = {
  key: string;
  data: TEventData;
  createdTxnHash: string;
  cancelledTxnHash?: string;
  executedTxnHash?: string;
  createdAt: number;
  isViewed?: boolean;
};

export type OrderStatus = MultiTransactionStatus<OrderCreatedEventData>;
// ...
export type PendingPositionUpdate = {
  isIncrease: boolean;
  positionKey: string;
  sizeDeltaUsd: BigNumber;
  sizeDeltaInTokens: BigNumber;
  collateralDeltaAmount: BigNumber;
  updatedAt: number;
  updatedAtBlock: BigNumber;
};